+++ system/modules/core/classes/Theme.php
@@ -315,21 +315,7 @@ class Theme extends \Backend
 				// Extract the files
 				try
 				{
-					$strFileName = $objArchive->file_name;
-
-					// Support the old "tl_files" directory
-					if (strncmp($strFileName, 'tl_files/', 9) === 0)
-					{
-						$strFileName = substr($strFileName, 3);
-					}
-
-					// Override the files directory
-					if ($GLOBALS['TL_CONFIG']['uploadPath'] != 'files' && strncmp($strFileName, 'files/', 6) === 0)
-					{
-						$strFileName = preg_replace('@^files/@', $GLOBALS['TL_CONFIG']['uploadPath'] . '/', $strFileName);
-					}
-
-					\File::putContent($strFileName, $objArchive->unzip());
+					\File::putContent($this->customizeUploadPath($objArchive->file_name), $objArchive->unzip());
 				}
 				catch (\Exception $e)
 				{
@@ -373,19 +359,7 @@ class Theme extends \Backend
 			{
 				foreach ($arrNewFolders as $strFolder)
 				{
-					// Support the old "tl_files" folder
-					if (strncmp($strFolder, 'tl_files/', 9) === 0)
-					{
-						$strFolder = substr($strFolder, 3);
-					}
-
-					// Override the files directory
-					if ($GLOBALS['TL_CONFIG']['uploadPath'] != 'files')
-					{
-						$strFolder = preg_replace('@^files/@', $GLOBALS['TL_CONFIG']['uploadPath'] . '/', $strFolder);
-					}
-
-					\Dbafs::addResource($strFolder);
+					\Dbafs::addResource($this->customizeUploadPath($strFolder));
 				}
 			}
 
@@ -445,12 +419,6 @@ class Theme extends \Backend
 					{
 						$value = $fields->item($k)->nodeValue;
 						$name = $fields->item($k)->getAttribute('name');
-
-						// Support the old "tl_files" folder
-						if (strncmp($value, 'tl_files/', 9) === 0)
-						{
-							$value = substr($value, 3);
-						}
 
 						// Skip NULL values
 						if ($value == 'NULL')
@@ -534,7 +502,7 @@ class Theme extends \Backend
 						}
 
 						// Adjust the file paths in style sheets
-						elseif ($GLOBALS['TL_CONFIG']['uploadPath'] != 'files' && ($table == 'tl_style_sheet' || $table == 'tl_style') && strpos($value, 'files') !== false)
+						elseif (($table == 'tl_style_sheet' || $table == 'tl_style') && strpos($value, 'files') !== false)
 						{
 							$tmp = deserialize($value);
 
@@ -542,14 +510,14 @@ class Theme extends \Backend
 							{
 								foreach ($tmp as $kk=>$vv)
 								{
-									$tmp[$kk] = preg_replace('@^files/@', $GLOBALS['TL_CONFIG']['uploadPath'] . '/', $vv);
+									$tmp[$kk] = $this->customizeUploadPath($vv);
 								}
 
 								$value = serialize($tmp);
 							}
 							else
 							{
-								$value = preg_replace('@^files/@', $GLOBALS['TL_CONFIG']['uploadPath'] . '/', $value);
+								$value = $this->customizeUploadPath($value);
 							}
 						}
 
@@ -565,7 +533,7 @@ class Theme extends \Backend
 								// Do not use the FilesModel here – tables are locked!
 								$objFile = $this->Database->prepare("SELECT uuid FROM tl_files WHERE path=?")
 														  ->limit(1)
-														  ->execute($value);
+														  ->execute($this->customizeUploadPath($value));
 
 								$value = $objFile->uuid;
 							}
@@ -580,16 +548,10 @@ class Theme extends \Backend
 							{
 								foreach ($tmp as $kk=>$vv)
 								{
-									// Support the old "tl_files" folder
-									if (strncmp($vv, 'tl_files/', 9) === 0)
-									{
-										$vv = substr($vv, 3);
-									}
-
 									// Do not use the FilesModel here – tables are locked!
 									$objFile = $this->Database->prepare("SELECT uuid FROM tl_files WHERE path=?")
 															  ->limit(1)
-															  ->execute($vv);
+															  ->execute($this->customizeUploadPath($vv));
 
 									$tmp[$kk] = $objFile->uuid;
 								}
@@ -892,15 +854,7 @@ class Theme extends \Backend
 
 				if ($objFile !== null)
 				{
-					// Standardize the upload path if it is not "files"
-					if ($GLOBALS['TL_CONFIG']['uploadPath'] != 'files')
-					{
-						$v = 'files/' . preg_replace('@^'.preg_quote($GLOBALS['TL_CONFIG']['uploadPath'], '@').'/@', '', $objFile->path);
-					}
-					else
-					{
-						$v = $objFile->path;
-					}
+					$v = $this->standardizeUploadPath($objFile->path);
 				}
 				else
 				{
@@ -919,28 +873,24 @@ class Theme extends \Backend
 
 					if ($objFiles !== null)
 					{
-						// Standardize the upload path if it is not "files"
-						if ($GLOBALS['TL_CONFIG']['uploadPath'] != 'files')
-						{
-							$arrTmp = array();
+						$arrTmp = array();
 
-							while ($objFiles->next())
-							{
-								$arrTmp[] = 'files/' . preg_replace('@^'.preg_quote($GLOBALS['TL_CONFIG']['uploadPath'], '@').'/@', '', $objFiles->path);
-							}
-
-							$v = serialize($arrTmp);
-						}
-						else
+						while ($objFiles->next())
 						{
-							$v = serialize($objFiles->fetchEach('path'));
+							$arrTmp[] = $this->standardizeUploadPath($objFiles->path);
 						}
+
+						$v = serialize($arrTmp);
 					}
 					else
 					{
 						$v = 'NULL';
 					}
 				}
+			}
+			elseif ($t == 'tl_style' && ($k == 'bgimage' || $k == 'liststyleimage'))
+			{
+				$v = $this->standardizeUploadPath($v);
 			}
 
 			$value = $xml->createTextNode($v);
@@ -1045,5 +995,37 @@ class Theme extends \Backend
 				$objArchive->addFile($strFolder .'/'. $strFile);
 			}
 		}
+	}
+
+
+	/**
+	 * Replace files/ with the custom upload folder name
+	 * @param string
+	 * @return string
+	 */
+	protected function customizeUploadPath($strPath)
+	{
+		if ($strPath == '')
+		{
+			return '';
+		}
+
+		return preg_replace('@^(tl_)?files/@', $GLOBALS['TL_CONFIG']['uploadPath'] . '/', $strPath);
+	}
+
+
+	/**
+	 * Replace a custom upload folder name with files/
+	 * @param string
+	 * @return string
+	 */
+	protected function standardizeUploadPath($strPath)
+	{
+		if ($strPath == '')
+		{
+			return '';
+		}
+
+		return preg_replace('@^' . preg_quote($GLOBALS['TL_CONFIG']['uploadPath'], '@') . '/@', 'files/', $strPath);
 	}
 }