+++ system/helper/functions.php
@@ -138,6 +138,8 @@ function show_help_message()
  */
 function die_nicely($strTemplate, $strFallback)
 {
+	header('Content-type: text/html; charset=utf-8');
+
 	if (file_exists(TL_ROOT . "/templates/$strTemplate.html5"))
 	{
 		include TL_ROOT . "/templates/$strTemplate.html5";