+++ system/modules/core/controllers/FrontendIndex.php
@@ -410,17 +410,16 @@ class FrontendIndex extends \Frontend
 		if ($expire !== null && (\Config::get('cacheMode') == 'both' || \Config::get('cacheMode') == 'browser'))
 		{
 			header('Cache-Control: public, max-age=' . ($expire - time()));
-			header('Expires: ' . gmdate('D, d M Y H:i:s', $expire) . ' GMT');
-			header('Last-Modified: ' . gmdate('D, d M Y H:i:s', time()) . ' GMT');
 			header('Pragma: public');
+			header('Last-Modified: ' . gmdate('D, d M Y H:i:s', time()) . ' GMT');
+			header('Expires: ' . gmdate('D, d M Y H:i:s', $expire) . ' GMT');
 		}
 		else
 		{
-			header('Cache-Control: no-cache');
-			header('Cache-Control: pre-check=0, post-check=0', false);
+			header('Cache-Control: no-store, no-cache, must-revalidate, post-check=0, pre-check=0');
+			header('Pragma: no-cache');
 			header('Last-Modified: ' . gmdate('D, d M Y H:i:s') . ' GMT');
 			header('Expires: Fri, 06 Jun 1975 15:10:00 GMT');
-			header('Pragma: no-cache');
 		}
 
 		echo $strBuffer;