+++ system/modules/core/dca/tl_member.php
@@ -388,12 +388,13 @@ $GLOBALS['TL_DCA']['tl_member'] = array
 		),
 		'lastLogin' => array
 		(
+			'label'                   => &$GLOBALS['TL_LANG']['MSC']['lastLogin'],
 			'eval'                    => array('rgxp'=>'datim'),
 			'sql'                     => "int(10) unsigned NOT NULL default '0'"
 		),
 		'currentLogin' => array
 		(
-			'label'                   => &$GLOBALS['TL_LANG']['MSC']['lastLogin'],
+			'label'                   => &$GLOBALS['TL_LANG']['MSC']['currentLogin'],
 			'sorting'                 => true,
 			'flag'                    => 6,
 			'eval'                    => array('rgxp'=>'datim'),