+++ system/modules/core/library/Contao/Widget.php
@@ -984,7 +984,7 @@ abstract class Widget extends \BaseTemplate
 						// Validate the date (see #5086)
 						try
 						{
-							new \Date($varInput);
+							new \Date($varInput, \Date::getNumericDateFormat());
 						}
 						catch (\OutOfBoundsException $e)
 						{
@@ -1012,7 +1012,7 @@ abstract class Widget extends \BaseTemplate
 						// Validate the date (see #5086)
 						try
 						{
-							new \Date($varInput);
+							new \Date($varInput, \Date::getNumericDatimFormat());
 						}
 						catch (\OutOfBoundsException $e)
 						{