+++ system/docs/CHANGELOG.md
@@ -5,6 +5,9 @@ Version 3.2.17 (2015-01-XX)
 ---------------------------
 
 ### Fixed
+Added two missing `exclude` flags in the `tl_page` data container (see #7522).
+
+### Fixed
 Send an UTF-8 charset header in the `die_nicely()` function (see #7519).
 
 ### Fixed